+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     - `AnVILConsortiumManagerEditRequired` -> `AnVILConsortiumManagerStaffEditRequired`
     - `AnVILConsortiumManagerViewRequired` -> `AnVILConsortiumManagerStaffViewRequired`
     - `AnVILConsortiumManagerLimitedViewRequired` -> `AnVILConsortiumManagerViewRequired`
+* Bugfix: Allow Workspace Data objects to have a second foreign key to `Workspace`.
 
 ## 0.19 (2023-10-27)